--- examples/example_worldspace/cache/packagejson.hash
@@ -1,1 +1,0 @@
-@types/react-reconciler:0.28.0@wonderlandengine/api:^1.1.0@wonderlandengine/components:^1.1.5@wonderlandengine/prettier-config:^1.0.0@wonderlandengine/react-ui:file:../../packages/react-uigl-matrix:^3.4.3@wonderlandengine/cli:^0.1.4
