+++ examples/example_worldspace/package.json
@@ -15,8 +15,8 @@
     ],
     "dependencies": {
         "@types/react-reconciler": "0.28.0",
-        "@wonderlandengine/api": "^1.1.0",
-        "@wonderlandengine/components": "^1.1.5",
+        "@wonderlandengine/api": "^1.4.6",
+        "@wonderlandengine/components": "^1.2.3",
         "@wonderlandengine/prettier-config": "^1.0.0",
         "@wonderlandengine/react-ui": "file:../../packages/react-ui",
         "gl-matrix": "^3.4.3"
